Upload files
This project was an extra credit assignment I did while attending Bemidji State University. I worked on this presentation at displayed it at the colleges Student Achievement Conference! My presentation was a look into the computer science degree. I created this project so that I could explain to my friends what programming is about. To demystify the major. This is also an effort to help the students decide if they like programming enough to take the first course! After all this presentation doubles as a guide to what is taught in the first classes. It was a good refresher of the basics and is also posted to research gate here: https://www.researchgate.net/publication/369825567_Introducing_Python_to_Non-CS_Students